backend/requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
pydantic==2.6.1
sqlalchemy==2.0.25
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
# For PostgreSQL (optional)
# psycopg2-binary==2.9.9 